Add x-dataselect to provide access to the native date picker.
This component uses `<input type="date">` with minor additions to align styling
with xel.

Styling alignment includes two `x-icon` elements to function as the clear and
select buttons.  They are over-layed on top of the native date controls and set
to allow all pointer events to pass to the underlying native buttons:
- the `date-picker` icon is over-layed on top of the native picker button
  (`-webkit-calendar-picker-indicator`);
- the `date-clear` icon is over-layed on top of the native clear button
  (`-webkit-clear-button`) *iff* the input is empty and the component
  *does not have* the `no-clear` attribute;
- the native buttons are sized to fit beneath the icons and are hidden with
  `opacity:0`;
- the native field wrapper (`-webkit-datetime-edit-fields-wrapper`) is hidden
  if the input is empty and not-active/focused to allow the xel label to be
  presented as the placeholder;
- the native spin-controls are currently disabled;